koma_app/requirements.txt
streamlit>=1.28.0
httpx>=0.25.0
xmltodict>=0.13.0
cachetools>=5.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
tenacity>=8.2.0xmltodict
